+++ Changes.md
@@ -10,6 +10,7 @@ Improvements
   - Increased the size of the triangle indicators for the decay ranges.
   - The decay range indicators are now scaled by the light's `gl:visualiser:scale` attribute.
   - The decay range is now ignored when framing a light in the Viewer.
+- EditScopePlugValueWidget : When viewing the output of a Box, the menu now displays the Edit Scopes contained within.
 
 API
 ---